# M10/tms/tms-api/contract/contract-types-orval/model/vehicleBrandListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * TMS API
 * Transportation Management System API
 * OpenAPI spec version: 1.0.0
 */
import type { Pagination } from './pagination';
import type { VehicleBrand } from './vehicleBrand';

export interface VehicleBrandListResponse {
  data: VehicleBrand[];
  pagination: Pagination;
}