# M10/tms/tms-api/contract/contract-types-orval/model/vehicleDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * TMS API
 * Transportation Management System API
 * OpenAPI spec version: 1.0.0
 */
import type { Vehicle } from './vehicle';
import type { VehicleBrand } from './vehicleBrand';
import type { VehicleDocument } from './vehicleDocument';
import type { VehicleHistoryEvent } from './vehicleHistoryEvent';
import type { VehicleModel } from './vehicleModel';

/**
 * A vehicle instance with its catalog (model + brand) and nested documents and history. Returned by GET /vehicles/{id}.

 */
export type VehicleDetail = Vehicle & {
  model?: VehicleModel;
  brand?: VehicleBrand;
  documents?: VehicleDocument[];
  history?: VehicleHistoryEvent[];
};