# M10/tms/tms-api/contract/contract-types-orval/model/vehicleKind.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * TMS API
 * Transportation Management System API
 * OpenAPI spec version: 1.0.0
 */

/**
 * Rozróżnienie ciągnik siodłowy vs naczepa.
 */
export type VehicleKind = typeof VehicleKind[keyof typeof VehicleKind];


export const VehicleKind = {
  TRACTOR_UNIT: 'TRACTOR_UNIT',
  SEMI_TRAILER: 'SEMI_TRAILER',
} as const;