# M10/tms/tms-api/contract/contract-types-orval/model/vehicleModel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * TMS API
 * Transportation Management System API
 * OpenAPI spec version: 1.0.0
 */
import type { VehicleKind } from './vehicleKind';
import type { VehicleModelTrailerType } from './vehicleModelTrailerType';

/**
 * A vehicle model (model pojazdu).
 */
export interface VehicleModel {
  id: number;
  brandId: number;
  /** @maxLength 120 */
  name: string;
  kind: VehicleKind;
  /**
   * Wymagane dla SEMI_TRAILER, puste dla TRACTOR_UNIT.
   * @nullable
   */
  trailerType?: VehicleModelTrailerType;
}