# M10/tms/tms-api/contract/openapi.yaml
openapi: 3.0.3
info:
  title: TMS API
  description: Transportation Management System API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local environment
...
paths:
...
  /vehicles:
    get:
      summary: List vehicles
      description: Returns a paginated list of all vehicles.
      operationId: getVehicles
      tags:
        - Vehicles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: kind
          in: query
          description: Filter by kind (TRACTOR_UNIT or SEMI_TRAILER)
          required: false
          schema:
            $ref: '#/components/schemas/VehicleKind'
        - name: modelId
          in: query
          description: Filter by catalog model id
          required: false
          schema:
            type: string
          example: '1'
      responses:
        '200':
          description: Paginated list of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
              example:
                data:
                  - id: 1
                    make: Suzuki
                    model: S4
                    year: 2021
                    fuel_tank_capacity: '51.1'
                  - id: 2
                    make: Volvo
                    model: S80 Fwd
                    year: 2022
                    fuel_tank_capacity: '52.2'
                pagination:
                  page: 1
                  limit: 20
                  total: 50
                  totalPages: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a vehicle
      description: Creates a new vehicle record.
      operationId: createVehicle
      tags:
        - Vehicles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreateInput'
            example:
              make: Toyota
              model: Corolla
              year: 2024
              fuel_tank_capacity: 55
      responses:
        '201':
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
              example:
                id: 51
                make: Toyota
                model: Corolla
                year: 2024
                fuel_tank_capacity: '55.0'
        '400':
          description: Request payload has an invalid structure or contains missing/invalid fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vehicles/{id}:
    get:
      summary: Get vehicle by ID
      description: Returns a single vehicle by its numeric ID.
      operationId: getVehicleById
      tags:
        - Vehicles
      parameters:
        - $ref: '#/components/parameters/PathIdParam'
      responses:
        '200':
          description: Vehicle found (with catalog, documents and history)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDetail'
        '400':
          description: The provided ID is missing or has an invalid format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No vehicle exists with the given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Vehicle not found
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update a vehicle
      description: Replaces vehicle data. All fields are optional – only provided fields are updated.
      operationId: updateVehicle
      tags:
        - Vehicles
      parameters:
        - $ref: '#/components/parameters/PathIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdateInput'
            example:
              make: Toyota
              model: Camry
              year: 2025
              fuel_tank_capacity: 60
      responses:
        '200':
          description: Vehicle updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
              example:
                id: 1
                make: Toyota
                model: Camry
                year: 2025
                fuel_tank_capacity: '60.0'
        '400':
          description: Request payload has an invalid structure or contains missing/invalid fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No vehicle exists with the given ID, or the update produced no result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Vehicle not found or invalid data
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a vehicle
      description: Permanently deletes a vehicle by its ID.
      operationId: deleteVehicle
      tags:
        - Vehicles
      parameters:
        - $ref: '#/components/parameters/PathIdParam'
      responses:
        '204':
          description: Vehicle deleted successfully – no body returned.
        '400':
          description: The provided ID is not a valid positive integer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No vehicle exists with the given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Vehicle not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /vehicle-brands:
    get:
      summary: List vehicle brands
      description: Returns a paginated list of vehicle brands (marki).
      operationId: getVehicleBrands
      tags:
        - VehicleBrands
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of vehicle brands
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleBrandListResponse'
              example:
                data:
                  - id: 1
                    name: Volvo
                    country: Sweden
                  - id: 2
                    name: Krone
                    country: Germany
                pagination:
                  page: 1
                  limit: 20
                  total: 12
                  totalPages: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a vehicle brand
      operationId: createVehicleBrand
      tags:
        - VehicleBrands
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleBrandCreateInput'
            example:
              name: Scania
              country: Sweden
      responses:
        '201':
          description: Vehicle brand created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleBrand'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vehicle-brands/{id}:
    get:
      summary: Get vehicle brand by ID
      operationId: getVehicleBrandById
      tags:
        - VehicleBrands
      parameters:
        - $ref: '#/components/parameters/PathIdParam'
      responses:
        '200':
          description: Vehicle brand found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleBrand'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update a vehicle brand
      operationId: updateVehicleBrand
      tags:
        - VehicleBrands
      parameters:
        - $ref: '#/components/parameters/PathIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleBrandUpdateInput'
            example:
              country: Sweden
      responses:
        '200':
          description: Vehicle brand updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleBrand'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a vehicle brand
      operationId: deleteVehicleBrand
      tags:
        - VehicleBrands
      parameters:
        - $ref: '#/components/parameters/PathIdParam'
      responses:
        '204':
          description: Vehicle brand deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      required: false
      schema:
        type: string
        default: '1'
      example: '1'
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (max 100)
      required: false
      schema:
        type: string
        default: '20'
      example: '20'
    PathIdParam:
      name: id
      in: path
      description: Numeric resource identifier
      required: true
      schema:
        type: integer
      example: 1
...
  responses:
    BadRequest:
      description: The request contains unknown or invalid query parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: An unexpected server-side error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The request conflicts with the current state (e.g. duplicate or referenced resource).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Pagination:
      type: object
      description: Pagination metadata attached to list responses.
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          description: Current page number (1-based)
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 20
        total:
          type: integer
          description: Total number of records across all pages
          example: 50
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
    ErrorResponse:
      type: object
      description: Standard error envelope returned on 4xx / 5xx responses.
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: Vehicle not found
...
    VehicleKind:
      type: string
      description: Rozróżnienie ciągnik siodłowy vs naczepa.
      enum:
        - TRACTOR_UNIT
        - SEMI_TRAILER
    Vehicle:
      type: object
      description: A single vehicle record as returned by the API.
      required:
        - id
        - model
      properties:
        id:
          type: integer
          description: Unique vehicle identifier
          example: 1
        make:
          type: string
          nullable: true
          maxLength: 50
          description: Vehicle manufacturer / brand
          example: Suzuki
        model:
          type: string
          maxLength: 50
          description: Vehicle model name
          example: S4
        year:
          type: integer
          nullable: true
          description: Manufacturing year
          example: 2021
        fuel_tank_capacity:
          type: string
          nullable: true
          description: |
            Fuel tank capacity in litres. Returned as a decimal string (e.g. "51.1") because Postgres numeric/decimal is serialised as a string by the pg driver.
          example: '51.1'
        model_id:
          type: integer
          nullable: true
          description: Reference to a catalog model (vehicle_models.id)
          example: 3
        kind:
          type: string
          nullable: true
          description: Rozróżnienie ciągnik siodłowy / naczepa (zdenormalizowane pod filtr)
          enum:
            - TRACTOR_UNIT
            - SEMI_TRAILER
        registration_number:
          type: string
          nullable: true
          maxLength: 20
          description: Numer rejestracyjny
          example: WA 12345
        vin:
          type: string
          nullable: true
          maxLength: 17
          description: Numer VIN
          example: YV2RT40A8FB123456
        first_registration_date:
          type: string
          format: date
          nullable: true
          description: Data pierwszej rejestracji
          example: '2021-03-15'
        mileage_km:
          type: integer
          nullable: true
          description: Przebieg w kilometrach
          example: 145000
        status:
          type: string
          nullable: true
          maxLength: 20
          description: Status egzemplarza
          example: active
        specs:
          type: object
          nullable: true
          additionalProperties: true
          description: |
            Techniczne atrybuty zależne od typu (JSONB).
            Ciągnik: {power_kw, euro_norm, axles, fuel_type};
            naczepa: {euro_pallets, volume_m3, interior_height_m, has_tail_lift, has_refrigeration}.
          example:
            power_kw: 331
            euro_norm: EURO6
            axles: 2
    VehicleListResponse:
      type: object
      description: Paginated list of vehicles.
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
        pagination:
          $ref: '#/components/schemas/Pagination'
    VehicleCreateInput:
      type: object
      description: |
        Payload for creating a new vehicle instance (egzemplarz). All fields are optional — use legacy make/model or link a catalog model via model_id.
      properties:
        make:
          type: string
          nullable: true
          maxLength: 50
          example: Volvo
        model:
          type: string
          nullable: true
          maxLength: 50
          example: FH 500
        year:
          type: integer
          nullable: true
          example: 2024
        fuel_tank_capacity:
          type: number
          format: float
          nullable: true
          minimum: 0
          exclusiveMinimum: true
          example: 600
        model_id:
          type: integer
          nullable: true
          example: 3
        kind:
          type: string
          nullable: true
          enum:
            - TRACTOR_UNIT
            - SEMI_TRAILER
        registration_number:
          type: string
          nullable: true
          maxLength: 20
          example: WA 12345
        vin:
          type: string
          nullable: true
          maxLength: 17
          example: YV2RT40A8FB123456
        first_registration_date:
          type: string
          format: date
          nullable: true
          example: '2021-03-15'
        mileage_km:
          type: integer
          nullable: true
          minimum: 0
          example: 145000
        status:
          type: string
          nullable: true
          maxLength: 20
          example: active
        specs:
          type: object
          nullable: true
          additionalProperties: true
          example:
            power_kw: 331
            euro_norm: EURO6
            axles: 2
    VehicleModel:
      type: object
      description: A vehicle model (model pojazdu).
      required:
        - id
        - brandId
        - name
        - kind
      properties:
        id:
          type: integer
          example: 1
        brandId:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 120
          example: Scania R450
        kind:
          $ref: '#/components/schemas/VehicleKind'
        trailerType:
          type: string
          nullable: true
          description: Wymagane dla SEMI_TRAILER, puste dla TRACTOR_UNIT.
          enum:
            - reefer
            - curtain
            - isotherm
            - tipper
            - platform
            - tank
            - container
    VehicleBrand:
      type: object
      description: A vehicle brand (marka).
      required:
        - id
        - name
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 80
          example: Volvo
        country:
          type: string
          nullable: true
          maxLength: 80
          example: Sweden
    VehicleDocument:
      type: object
      description: A document (dokument) attached to a vehicle.
      required:
        - id
        - vehicle_id
        - doc_type
      properties:
        id:
          type: integer
          example: 1
        vehicle_id:
          type: integer
          example: 3
        doc_type:
          type: string
          description: registration_certificate | insurance_oc | insurance_ac | technical_inspection | tachograph | atp_certificate | other
          example: insurance_oc
        document_number:
          type: string
          nullable: true
          maxLength: 60
          example: OC/2026/12345
        issue_date:
          type: string
          format: date
          nullable: true
          example: '2026-01-01'
        expiry_date:
          type: string
          format: date
          nullable: true
          example: '2026-12-31'
        file_url:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    VehicleHistoryEvent:
      type: object
      description: A single history event (zdarzenie historii) of a vehicle.
      required:
        - id
        - vehicle_id
        - event_type
        - event_date
      properties:
        id:
          type: integer
          example: 1
        vehicle_id:
          type: integer
          example: 3
        event_type:
          type: string
          description: purchase | inspection | repair | accident | mileage_reading | status_change | sold
          example: inspection
        event_date:
          type: string
          format: date
          example: '2026-05-10'
        mileage_km:
          type: integer
          nullable: true
          example: 145000
        description:
          type: string
          nullable: true
          example: Badanie techniczne — wynik pozytywny
    VehicleDetail:
      description: |
        A vehicle instance with its catalog (model + brand) and nested documents and history. Returned by GET /vehicles/{id}.
      allOf:
        - $ref: '#/components/schemas/Vehicle'
        - type: object
          properties:
            model:
              $ref: '#/components/schemas/VehicleModel'
            brand:
              $ref: '#/components/schemas/VehicleBrand'
            documents:
              type: array
              items:
                $ref: '#/components/schemas/VehicleDocument'
            history:
              type: array
              items:
                $ref: '#/components/schemas/VehicleHistoryEvent'
    VehicleUpdateInput:
      allOf:
        - $ref: '#/components/schemas/VehicleCreateInput'
      description: |
        Payload for updating an existing vehicle. All fields are optional (partial update). Provide only the fields you want to change.
...
    VehicleBrandListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/VehicleBrand'
        pagination:
          $ref: '#/components/schemas/Pagination'
    VehicleBrandCreateInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          example: Scania
        country:
          type: string
          nullable: true
          maxLength: 80
          example: Sweden
    VehicleBrandUpdateInput:
      type: object
      description: Partial update — only provided fields are changed.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        country:
          type: string
          nullable: true
          maxLength: 80
...